Relocate TimeSeriesDataTable class to a new file.
